harness: absorb recurrent into substrate
