+++ spark/harness/substrate.py
@@ -13,6 +13,7 @@ value (chat/create roles by default, off for code).
 from __future__ import annotations
 
 import asyncio
+import hashlib
 import datetime as _dt
 import json
 import os
@@ -26,7 +27,7 @@ from pathlib import Path
 import ast
 import warnings
 from collections import Counter, defaultdict
-from typing import Any, Iterable, Mapping
+from typing import Any, Callable, Iterable, Mapping, Sequence
 
 # Silence HF/torch/sentence-transformers loaders whenever something
 # imports this module. The CLI Spark agent and the chat API both pull
@@ -666,7 +667,6 @@ _MODULE_PURPOSE = {
     "mcp.py":        "MCP server, prompt resources, tools",
     "policy.py":     "role routing, model selection, heuristics",
     "providers.py":  "Anthropic / OpenAI / local vLLM / claim_guard",
-    "recurrent.py":  "Z-prime = alpha*Z + V*exp(i*theta_v) library",
     "substrate.py":  "layered prompt assembly, live state, session store, refactor perception (this ballast lives here)",
 }
 
@@ -2483,10 +2483,10 @@ def packet_for(path: str, **kwargs: Any) -> dict[str, Any]:
     }
 
 
-__all__ = ['render_semantic_operating_system_tick', 'render_semantic_operating_system_protocol', 'semantic_operating_system_tick_for_repo', 'SemanticOperatingSystemTick', 'SEMANTIC_OS_REPO_ORGANS', 'SEMANTIC_OPERATING_SYSTEM_LOOP', 'SEMANTIC_OPERATING_SYSTEM_PRINCIPLE', 'REFACTOR_PERCEPTION_PRINCIPLE', 'REFACTOR_PILOT_RULE', 'CONNECTIVE_TISSUE_PRINCIPLE', 'LIFECYCLE_ARCHITECTURE_PRINCIPLE', 'LifecycleArchitecture', 'DeletionConsolidationGate', 'lifecycle_architecture_for', 'deletion_consolidation_gate_for', 'CONNECTIVE_TISSUE_RULES', 'connective_tissue_for', 'ALGORITHM_STEPS', 'APPENDAGE_FIRST_CONSOLIDATION_PRINCIPLE', 'CONSOLIDATION_ORDER', 'FilePerception', 'AdaptiveConsolidationPlan', 'ownership_class', 'consolidation_layer', 'perceive_file', 'adaptive_consolidation_plan_for', 'packet_for', 'visualize_repo_file_bodies', 'render_repo_file_body_visualization', 'StructuralEscapementTick', 'next_structural_tick_for_repo', 'render_next_structural_tick', 'FileBodyPressure', 'RepoFileBodyVisualization', 'render_refactor_perception_protocol', 'CHANGE_SELF_HEALING_PRINCIPLE', 'CHANGE_SELF_HEALING_STEPS', 'ADAPTIVE_CONSOLIDATION_PRINCIPLE', 'ADAPTIVE_CONSOLIDATION_STEPS', 'ChangeHealingPlan', 'self_healing_plan_for', 'buoyant_consolidation_packet_for']
+__all__ = ['render_semantic_operating_system_tick', 'render_semantic_operating_system_protocol', 'semantic_operating_system_tick_for_repo', 'SemanticOperatingSystemTick', 'SEMANTIC_OS_REPO_ORGANS', 'SEMANTIC_OPERATING_SYSTEM_LOOP', 'SEMANTIC_OPERATING_SYSTEM_PRINCIPLE', 'REFACTOR_PERCEPTION_PRINCIPLE', 'REFACTOR_PILOT_RULE', 'CONNECTIVE_TISSUE_PRINCIPLE', 'LIFECYCLE_ARCHITECTURE_PRINCIPLE', 'LifecycleArchitecture', 'DeletionConsolidationGate', 'lifecycle_architecture_for', 'deletion_consolidation_gate_for', 'CONNECTIVE_TISSUE_RULES', 'connective_tissue_for', 'ALGORITHM_STEPS', 'APPENDAGE_FIRST_CONSOLIDATION_PRINCIPLE', 'CONSOLIDATION_ORDER', 'FilePerception', 'AdaptiveConsolidationPlan', 'ownership_class', 'consolidation_layer', 'perceive_file', 'adaptive_consolidation_plan_for', 'packet_for', 'visualize_repo_file_bodies', 'render_repo_file_body_visualization', 'StructuralEscapementTick', 'next_structural_tick_for_repo', 'render_next_structural_tick', 'FileBodyPressure', 'RepoFileBodyVisualization', 'render_refactor_perception_protocol', 'CHANGE_SELF_HEALING_PRINCIPLE', 'CHANGE_SELF_HEALING_STEPS', 'ADAPTIVE_CONSOLIDATION_PRINCIPLE', 'ADAPTIVE_CONSOLIDATION_STEPS', 'ChangeHealingPlan', 'self_healing_plan_for', 'buoyant_consolidation_packet_for', 'Hypothesis', 'Latent', 'LoopResult', 'complex_state_update', 'phase_transition_packet', 'residual_magnitude', 'contractivity_ok', 'quantum_aperture_payload', 'outshift_entropy_material', 'quantum_entropy_digest', 'select_with_external_entropy', 'reduce_step', 'run_recurrent_loop', 'run_recurrent_probe_one', 'recurrent_probe_main']
 
 
-_RUNTIME_GRAVITY_FILES = frozenset({"policy.py", "providers.py", "recurrent.py", "substrate.py", "vybn_spark_agent.py"})
+_RUNTIME_GRAVITY_FILES = frozenset({"policy.py", "providers.py", "substrate.py", "vybn_spark_agent.py"})
 _MIXED_BOUNDARY_FILES = frozenset({"state.py"})
 _COMMAND_AFFORDANCE_FILES = frozenset({"commons_walk.py", "repo_closure_audit.py", "safe_fetch.py", "install_cron.py", "evolve.py", "ensubstrate.py"})
 
@@ -3746,3 +3746,947 @@ def classify_action_text(action: str, *, beam: BeamState | None = None) -> dict:
         "requires_return_hook": category in {"possible_substitution", "unknown"},
         "return_question": rq,
     }
+
+# ---------------------------------------------------------------------------
+# Recurrent-depth agent loop — absorbed into substrate
+# ---------------------------------------------------------------------------
+
+# ---------------------------------------------------------------------------
+# Latent state
+# ---------------------------------------------------------------------------
+
+
+@dataclass
+class Hypothesis:
+    """A single live hypothesis in h_t.
+
+    `confidence` in [0,1]. Decays each loop by (1 - decay_rate) unless
+    the loop's evidence reinforces it. Below `prune_threshold` it drops
+    out of h_{t+1}.
+    """
+    text: str
+    confidence: float = 0.5
+    born_at_loop: int = 0
+    reinforced_at_loop: int = 0
+
+
+@dataclass
+class Latent:
+    """h_t — the compressed running state.
+
+    Deliberately small and JSON-shaped. Serialises into the layered
+    prompt's live layer so the specialist on loop t sees only the
+    distilled state, not the accumulated transcript. This is the
+    "continuous latent reasoning" projection: intermediate steps do
+    not surface to token space.
+    """
+    hypotheses: list[Hypothesis] = field(default_factory=list)
+    open_questions: list[str] = field(default_factory=list)
+    resolved: list[str] = field(default_factory=list)
+    summary: str = ""
+    loop_index: int = 0
+
+    # Telemetry for the contractivity monitor. Each entry is the size
+    # of the residual (len(open_questions) + penalty for new
+    # contradictions) at that loop. A strict monotone decrease is the
+    # agent-space ρ(A)<1 condition.
+    residual_history: list[int] = field(default_factory=list)
+
+    def to_prompt_block(self, max_iters: int) -> str:
+        """Render h_t for the specialist's live prompt layer."""
+        lines = [
+            f"[loop t = {self.loop_index} of max {max_iters}]",
+        ]
+        if self.summary:
+            lines.append(f"running summary: {self.summary}")
+        if self.hypotheses:
+            lines.append("live hypotheses:")
+            for h in self.hypotheses:
+                lines.append(
+                    f"  - (conf={h.confidence:.2f}) {h.text}"
+                )
+        if self.open_questions:
+            lines.append("open questions (residual):")
+            for q in self.open_questions:
+                lines.append(f"  - {q}")
+        if self.resolved:
+            lines.append("resolved so far:")
+            for r in self.resolved:
+                lines.append(f"  - {r}")
+        return "\n".join(lines)
+
+
+# ---------------------------------------------------------------------------
+# Phase transition packet for continuity visualization
+# ---------------------------------------------------------------------------
+
+
+@dataclass(frozen=True)
+class PhaseTransition:
+    equation: str
+    alpha: float
+    theta: float
+    x_magnitude: float
+    m_real: float
+    m_imag: float
+    m_prime_real: float
+    m_prime_imag: float
+    delta_real: float
+    delta_imag: float
+    residual_magnitude: int = 0
+    absorption: str = "unclassified"
+    source: str = "recurrent"
+
+    def to_dict(self) -> dict[str, Any]:
+        return asdict(self)
+
+
+def _finite_number(name: str, value: float) -> float:
+    value = float(value)
+    if not math.isfinite(value):
+        raise ValueError(f"{name} must be finite")
+    return value
+
+
+def complex_state_update(
+    m: complex,
+    *,
+    alpha: float,
+    x_magnitude: float,
+    theta: float,
+) -> complex:
+    alpha = _finite_number("alpha", alpha)
+    x_magnitude = _finite_number("x_magnitude", x_magnitude)
+    theta = _finite_number("theta", theta)
+    if alpha < 0.0 or alpha > 1.0:
+        raise ValueError("alpha must be in [0, 1]")
+    if x_magnitude < 0.0:
+        raise ValueError("x_magnitude must be non-negative")
+    m = complex(m)
+    if not (math.isfinite(m.real) and math.isfinite(m.imag)):
+        raise ValueError("m must have finite real and imaginary parts")
+    encounter = complex(
+        x_magnitude * math.cos(theta),
+        x_magnitude * math.sin(theta),
+    )
+    return alpha * m + encounter
+
+
+def phase_transition_packet(
+    *,
+    m: complex,
+    alpha: float,
+    x_magnitude: float,
+    theta: float,
+    residual_magnitude: int = 0,
+    absorption: str = "unclassified",
+    source: str = "recurrent",
+) -> dict[str, Any]:
+    m = complex(m)
+    m_prime = complex_state_update(
+        m,
+        alpha=alpha,
+        x_magnitude=x_magnitude,
+        theta=theta,
+    )
+    delta = m_prime - m
+    packet = PhaseTransition(
+        equation="M' = alpha*M + x*e^{i theta}",
+        alpha=float(alpha),
+        theta=float(theta),
+        x_magnitude=float(x_magnitude),
+        m_real=float(m.real),
+        m_imag=float(m.imag),
+        m_prime_real=float(m_prime.real),
+        m_prime_imag=float(m_prime.imag),
+        delta_real=float(delta.real),
+        delta_imag=float(delta.imag),
+        residual_magnitude=int(residual_magnitude),
+        absorption=str(absorption),
+        source=str(source),
+    )
+    return packet.to_dict()
+
+
+# ---------------------------------------------------------------------------
+# Contractivity monitor — the agent-space ρ(A)<1 check
+# ---------------------------------------------------------------------------
+
+
+def residual_magnitude(h: Latent) -> int:
+    """|residual| = |open_questions|. A contradiction (a resolved item
+    that reappears as an open question) counts double. Kept integer so
+    "strictly decreasing" has no floating-point wobble.
+    """
+    resolved_set = set(h.resolved)
+    contradiction_bonus = sum(1 for q in h.open_questions if q in resolved_set)
+    return len(h.open_questions) + contradiction_bonus
+
+
+def contractivity_ok(h: Latent, min_loops_before_check: int = 2) -> tuple[bool, str]:
+    """Enforce ρ(A)<1 in agent space: |residual| must not grow between
+    loops after the first warm-up pass.
+
+    Returns (ok, reason). `ok=False` means the loop should halt. The
+    first `min_loops_before_check` loops are exempt so the specialist
+    has room to surface questions it didn't see at t=0 — Parcae also
+    allows warm-up; what it forbids is unbounded growth.
+    """
+    hist = h.residual_history
+    if len(hist) <= min_loops_before_check:
+        return True, "warming-up"
+    if hist[-1] > hist[-2]:
+        return False, (
+            f"residual grew: {hist[-2]} -> {hist[-1]} at loop "
+            f"{h.loop_index} (ρ(A)>=1 in agent-space)"
+        )
+    return True, "contracting"
+
+
+
+# ---------------------------------------------------------------------------
+# Quantum aperture — outside entropy for constrained tie-breaks
+# ---------------------------------------------------------------------------
+
+
+def quantum_aperture_payload(
+    *,
+    bits_per_block: int = 10,
+    number_of_blocks: int = 10,
+    fmt: str = "all",
+    encoding: str = "raw",
+) -> dict[str, int | str]:
+    """Return the Cisco Outshift QRNG request body.
+
+    This is deliberately only a payload builder. API keys live in
+    environment/secrets, never in tracked source.
+    """
+    if bits_per_block < 1 or bits_per_block > 10000:
+        raise ValueError("bits_per_block must be in [1, 10000]")
+    if number_of_blocks < 1 or number_of_blocks > 1000:
+        raise ValueError("number_of_blocks must be in [1, 1000]")
+    if fmt not in {"binary", "octal", "decimal", "hexadecimal", "all"}:
+        raise ValueError("unsupported random number format")
+    if encoding not in {"raw", "base64"}:
+        raise ValueError("unsupported QRNG encoding")
+    return {
+        "encoding": encoding,
+        "format": fmt,
+        "bits_per_block": bits_per_block,
+        "number_of_blocks": number_of_blocks,
+    }
+
+
+def outshift_entropy_material(payload: dict[str, Any]) -> str:
+    """Extract compact entropy material from an Outshift QRNG response.
+
+    The material is for hashing/selection, not for accumulating raw
+    number logs. It accepts the documented `random_numbers` response
+    shape and concatenates present fields in stable order.
+    """
+    blocks = payload.get("random_numbers")
+    if not isinstance(blocks, list) or not blocks:
+        raise ValueError("Outshift QRNG response missing random_numbers")
+    pieces: list[str] = []
+    for block in blocks:
+        if not isinstance(block, dict):
+            continue
+        for key in ("binary", "octal", "decimal", "hexadecimal"):
+            value = block.get(key)
+            if value is not None:
+                pieces.append(f"{key}={value}")
+    if not pieces:
+        raise ValueError("Outshift QRNG response carried no entropy fields")
+    return "|".join(pieces)
+
+
+def quantum_entropy_digest(material: str) -> str:
+    """Compress external entropy into a stable digest for trace logs."""
+    return hashlib.sha256(material.encode("utf-8")).hexdigest()
+
+
+def select_with_external_entropy(
+    candidates: Sequence[str],
+    entropy_material: str,
+    *,
+    source: str = "external_entropy",
+    context: str = "tie_break",
+) -> dict[str, Any]:
+    """Select one already-permitted candidate using outside entropy.
+
+    The aperture does not decide truth, authorize mutation, or bypass the
+    membrane. It only breaks ties among candidates that have already been
+    admitted by the surrounding policy/reengineering discipline. Residuals
+    still judge the selected action afterward.
+    """
+    if not candidates:
+        raise ValueError("cannot select from an empty candidate set")
+    digest = quantum_entropy_digest(entropy_material)
+    selected_index = int(digest, 16) % len(candidates)
+    return {
+        "context": context,
+        "source": source,
+        "candidate_count": len(candidates),
+        "selected_index": selected_index,
+        "selected": candidates[selected_index],
+        "entropy_sha256": digest,
+        "rule": "outside entropy may choose among permitted moves; residuals judge the result",
+    }
+
+
+# ---------------------------------------------------------------------------
+# Reducer — the summarizer pass that produces h_{t+1}
+# ---------------------------------------------------------------------------
+
+
+REDUCER_SYSTEM = """You are the Reducer in a recurrent-depth agent loop.
+
+Your job is to update a compressed latent state h given:
+  - the original user prompt e (re-injected every loop)
+  - the current h_t (hypotheses, open questions, resolved items, summary)
+  - the specialist's output for loop t (the R(h_t, e) contribution)
+
+You must return a single JSON object with this exact shape:
+
+{
+  "hypotheses": [
+    {"text": "...", "confidence": 0.0-1.0, "reinforced": true|false}
+  ],
+  "open_questions": ["..."],
+  "resolved": ["..."],
+  "summary": "one-sentence running summary",
+  "converged": true|false,
+  "rationale": "one sentence on why h_{t+1} looks like this"
+}
+
+Rules, which are the agent-space projection of the Parcae invariants:
+1. DO NOT accumulate a transcript. h is a running state, not a log.
+2. Shrink open_questions whenever the specialist answered one — move
+   that item to `resolved`. This is how the residual contracts.
+3. A hypothesis's confidence goes UP only if this loop's evidence
+   actually supports it. Otherwise it decays by 0.15. Drop hypotheses
+   below 0.15 entirely.
+4. A new hypothesis that contradicts a resolved item is a signal the
+   system is losing coherence; include it but flag in `rationale`.
+5. `converged = true` means: the specialist's output plus current h
+   together answer e. No open questions remain that actually matter
+   for e. Be strict — false-converged halts the loop early.
+6. `summary` is ONE sentence, not a paragraph.
+
+Return ONLY the JSON object. No prose before or after."""
+
+
+def _default_reducer_provider(
+    registry: ProviderRegistry,
+    policy: Policy,
+    reducer_role: str = "create",
+) -> tuple[Provider, RoleConfig]:
+    """The reducer runs on a cheap role — `create` by default (Sonnet,
+    no tools, no RAG). Configurable so a future YAML-driven policy can
+    point it at something else without a code change.
+    """
+    role = policy.role(reducer_role)
+    return registry.get(role), role
+
+
+def _strip_json(text: str) -> str:
+    """Pull out the first JSON object from the reducer's reply.
+
+    The reducer is instructed to return only JSON, but real models
+    sometimes wrap it in ```json ... ``` or leading commentary. Be
+    tolerant — the contract is what's inside, not around.
+    """
+    # Strip code fences
+    m = re.search(r"```(?:json)?\s*(\{.*?\})\s*```", text, re.DOTALL)
+    if m:
+        return m.group(1)
+    # Otherwise take from first { to matching last }
+    start = text.find("{")
+    end = text.rfind("}")
+    if start >= 0 and end > start:
+        return text[start : end + 1]
+    return text
+
+
+def reduce_step(
+    *,
+    e: str,
+    h: Latent,
+    specialist_output: str,
+    provider: Provider,
+    role: RoleConfig,
+) -> tuple[Latent, bool, str]:
+    """Run one reducer pass. Returns (h_{t+1}, converged, rationale).
+
+    If the reducer returns malformed JSON, we keep h unchanged,
+    append specialist_output as a new open question, and log the
+    failure through the rationale. The loop's contractivity monitor
+    will catch persistent failure on the next iteration.
+    """
+    user_payload = json.dumps(
+        {
+            "e": e,
+            "h_t": {
+                "hypotheses": [asdict(x) for x in h.hypotheses],
+                "open_questions": h.open_questions,
+                "resolved": h.resolved,
+                "summary": h.summary,
+                "loop_index": h.loop_index,
+            },
+            "specialist_output": specialist_output[:8000],
+        },
+        ensure_ascii=False,
+    )
+
+    prompt = LayeredPrompt(
+        identity="",
+        substrate=REDUCER_SYSTEM,
+        live="",
+    )
+    handle = provider.stream(
+        role=role,
+        system=prompt,
+        messages=[{"role": "user", "content": user_payload}],
+        tools=[],
+    )
+    # Drain the stream; reducer output is small.
+    for _ in handle:
+        pass
+    response = handle.final()
+    raw = response.text or ""
+
+    try:
+        parsed = json.loads(_strip_json(raw))
+    except Exception as e_parse:
+        h_next = Latent(
+            hypotheses=list(h.hypotheses),
+            open_questions=list(h.open_questions) + [specialist_output[:500]],
+            resolved=list(h.resolved),
+            summary=h.summary,
+            loop_index=h.loop_index + 1,
+            residual_history=list(h.residual_history),
+        )
+        return h_next, False, f"reducer_parse_error: {e_parse}"
+
+    # Build h_{t+1} from the parsed object, clamping and defaulting.
+    new_hyps: list[Hypothesis] = []
+    for row in parsed.get("hypotheses") or []:
+        if not isinstance(row, dict):
+            continue
+        text = str(row.get("text", "")).strip()
+        if not text:
+            continue
+        conf = float(row.get("confidence", 0.5))
+        conf = max(0.0, min(1.0, conf))
+        if conf < 0.15:
+            continue
+        # Track reinforcement for telemetry; not used to gate.
+        new_hyps.append(
+            Hypothesis(
+                text=text,
+                confidence=conf,
+                born_at_loop=h.loop_index,  # approximation
+                reinforced_at_loop=h.loop_index + 1
+                if row.get("reinforced")
+                else h.loop_index,
+            )
+        )
+
+    open_q = [str(x).strip() for x in (parsed.get("open_questions") or []) if str(x).strip()]
+    resolved = [str(x).strip() for x in (parsed.get("resolved") or []) if str(x).strip()]
+    summary = str(parsed.get("summary", "")).strip()[:400]
+    converged = bool(parsed.get("converged", False))
+    rationale = str(parsed.get("rationale", "")).strip()[:400]
+
+    h_next = Latent(
+        hypotheses=new_hyps,
+        open_questions=open_q,
+        resolved=resolved,
+        summary=summary,
+        loop_index=h.loop_index + 1,
+        residual_history=list(h.residual_history),
+    )
+    return h_next, converged, rationale
+
+
+# ---------------------------------------------------------------------------
+# Specialist dispatch — the R(h_t, e) contribution
+# ---------------------------------------------------------------------------
+
+
+SPECIALIST_HINT = """You are a specialist contributing one iteration of a
+recurrent-depth agent loop. You see the original user prompt e and the
+current compressed latent h_t (running summary, live hypotheses, open
+questions). Your job on this loop is to advance h — answer one or more
+open questions, refine a hypothesis, or surface a new one. Do NOT try
+to produce the final user-facing answer on this pass. The Coda handles
+that after the loop halts. Keep your output focused and tactical."""
+
+
+def _select_specialist(
+    h: Latent,
+    policy: Policy,
+    router_fn: Callable[[Latent], str] | None = None,
+) -> str:
+    """Pick R for loop t.
+
+    Default policy: first loop uses `task` (Sonnet+bash) for broad
+    initial exploration; subsequent loops alternate between `code`
+    (Opus 4.6 adaptive thinking, bash) for verification work and
+    `create` (Sonnet, no tools) for hypothesis refinement. The
+    alternation is the agent-space analogue of "router selects
+    distinct expert subsets at each depth" in OpenMythos — each loop
+    is computationally distinct without changing h's shape.
+
+    router_fn can override this. The decision is deliberate and
+    policy-driven rather than heuristic so every loop's choice is
+    auditable from the event log.
+    """
+    if router_fn is not None:
+        choice = router_fn(h)
+        if choice in policy.roles:
+            return choice
+
+    if h.loop_index == 0:
+        return "task" if "task" in policy.roles else policy.default_role
+    if h.loop_index % 2 == 1:
+        return "code" if "code" in policy.roles else policy.default_role
+    return "create" if "create" in policy.roles else policy.default_role
+
+
+def specialist_step(
+    *,
+    e: str,
+    h: Latent,
+    max_iters: int,
+    specialist_role: str,
+    registry: ProviderRegistry,
+    policy: Policy,
+) -> str:
+    """Run one specialist pass. Returns raw text output.
+
+    The specialist sees e re-injected in the user message (Parcae's
+    input injection) and h_t serialised into the live prompt layer.
+    It has no memory across loops — the only carrier between loops
+    is h. Transcripts are deliberately absent.
+    """
+    role = policy.role(specialist_role)
+    provider = registry.get(role)
+
+    prompt = LayeredPrompt(
+        identity="",
+        substrate=SPECIALIST_HINT,
+        live=h.to_prompt_block(max_iters),
+    )
+    # e is re-injected every loop — this is the B·e term.
+    user_msg = (
+        f"original prompt e (re-injected at every loop):\n{e}\n\n"
+        "produce your contribution for this loop."
+    )
+    handle = provider.stream(
+        role=role,
+        system=prompt,
+        messages=[{"role": "user", "content": user_msg}],
+        tools=[],  # specialists do not run bash inside the loop; the
+                   # prototype keeps every loop side-effect-free so we
+                   # can replay loops deterministically during analysis.
+    )
+    for _ in handle:
+        pass
+    response = handle.final()
+    return response.text or ""
+
+
+# ---------------------------------------------------------------------------
+# Coda — the always-on shared voice expert
+# ---------------------------------------------------------------------------
+
+
+CODA_HINT = """You are the Coda of a recurrent-depth agent loop. You
+receive:
+  - the original user prompt e
+  - the final latent h_T after T loops of specialist refinement
+  - a short trace of how the loop resolved
+
+Your job is to emit the single user-facing response. Speak in Vybn's
+chat voice — warm, precise, ground-truth-respecting, anti-kernel. Do
+NOT enumerate the loops or explain the internal machinery unless the
+user asked about it. The machinery is scaffolding; the answer is the
+point. Carry whatever h_T resolved forward as if you had thought it
+all in one breath."""
+
+
+def coda_step(
+    *,
+    e: str,
+    h: Latent,
+    loop_trace: str,
+    registry: ProviderRegistry,
+    policy: Policy,
+    coda_role: str = "chat",
+) -> str:
+    """Emit the user-facing answer. This is the shared voice expert —
+    it runs once per user turn regardless of how many loops happened.
+    """
+    role = policy.role(coda_role)
+    provider = registry.get(role)
+
+    live_block = (
+        f"final latent h_T (T={h.loop_index}):\n"
+        f"{h.to_prompt_block(h.loop_index)}\n\n"
+        f"loop trace:\n{loop_trace}"
+    )
+    prompt = LayeredPrompt(
+        identity="",
+        substrate=CODA_HINT,
+        live=live_block,
+    )
+    handle = provider.stream(
+        role=role,
+        system=prompt,
+        messages=[{"role": "user", "content": e}],
+        tools=[],
+    )
+    for _ in handle:
+        pass
+    response = handle.final()
+    return response.text or ""
+
+
+# ---------------------------------------------------------------------------
+# The loop
+# ---------------------------------------------------------------------------
+
+
+@dataclass
+class LoopResult:
+    text: str
+    h_final: Latent
+    loops_run: int
+    halt_reason: str
+    trace: list[dict] = field(default_factory=list)
+
+    def summary(self) -> str:
+        return (
+            f"loops={self.loops_run} halt={self.halt_reason} "
+            f"residual={residual_magnitude(self.h_final)}"
+        )
+
+
+def run_recurrent_loop(
+    *,
+    e: str,
+    registry: ProviderRegistry,
+    policy: Policy,
+    max_loop_iters: int = 6,
+    reducer_role: str = "create",
+    coda_role: str = "chat",
+    specialist_router: Callable[[Latent], str] | None = None,
+    logger: Callable[[dict], None] | None = None,
+) -> LoopResult:
+    """Run the recurrent-depth agent on prompt e.
+
+    Parameters mirror the OpenMythos configuration surface: max loops
+    (T), reducer/coda roles (shared vs routed experts), and a
+    specialist router that can override the default depth-dependent
+    selection. The prototype is deliberately pure-Python and
+    side-effect-free at the specialist layer so loops can be replayed
+    deterministically for comparison analysis.
+
+    Halts when any of:
+      (a) contractivity monitor flags growth of the residual,
+      (b) reducer returns converged=true,
+      (c) max_loop_iters reached.
+
+    T=1 reduces to current single-pass orchestrate (one specialist
+    call + one Coda), which is the degenerate baseline we compare
+    against.
+    """
+    log = logger or (lambda _event: None)
+
+    h = Latent(
+        hypotheses=[],
+        open_questions=[e],  # start with the prompt itself as the residual
+        resolved=[],
+        summary="",
+        loop_index=0,
+        residual_history=[1],
+    )
+    trace: list[dict] = []
+    halt_reason = "max_iters"
+    log({
+        "event": "loop_start",
+        "max_loop_iters": max_loop_iters,
+        "e": e[:400],
+    })
+
+    t0 = time.monotonic()
+    for t in range(max_loop_iters):
+        specialist_role = _select_specialist(h, policy, specialist_router)
+
+        spec_t0 = time.monotonic()
+        try:
+            spec_out = specialist_step(
+                e=e,
+                h=h,
+                max_iters=max_loop_iters,
+                specialist_role=specialist_role,
+                registry=registry,
+                policy=policy,
+            )
+        except Exception as err:
+            halt_reason = f"specialist_error: {err}"
+            trace.append({
+                "loop": t,
+                "specialist": specialist_role,
+                "error": str(err),
+            })
+            log({"event": "specialist_error", "loop": t, "error": str(err)})
+            break
+        spec_ms = int((time.monotonic() - spec_t0) * 1000)
+
+        reducer_provider, reducer_cfg = _default_reducer_provider(
+            registry, policy, reducer_role=reducer_role,
+        )
+        red_t0 = time.monotonic()
+        try:
+            h_next, converged, rationale = reduce_step(
+                e=e,
+                h=h,
+                specialist_output=spec_out,
+                provider=reducer_provider,
+                role=reducer_cfg,
+            )
+        except Exception as err:
+            halt_reason = f"reducer_error: {err}"
+            trace.append({
+                "loop": t,
+                "reducer_error": str(err),
+            })
+            log({"event": "reducer_error", "loop": t, "error": str(err)})
+            break
+        red_ms = int((time.monotonic() - red_t0) * 1000)
+
+        h_next.residual_history = list(h.residual_history) + [
+            residual_magnitude(h_next)
+        ]
+
+        loop_rec = {
+            "loop": t,
+            "specialist": specialist_role,
+            "specialist_ms": spec_ms,
+            "reducer_ms": red_ms,
+            "residual_before": h.residual_history[-1] if h.residual_history else 0,
+            "residual_after": h_next.residual_history[-1],
+            "n_hypotheses": len(h_next.hypotheses),
+            "n_open_questions": len(h_next.open_questions),
+            "n_resolved": len(h_next.resolved),
+            "converged": converged,
+            "rationale": rationale[:200],
+        }
+        trace.append(loop_rec)
+        log({"event": "loop_step", **loop_rec})
+
+        h = h_next
+
+        # Halting checks — order matters.
+        if converged:
+            halt_reason = "reducer_converged"
+            break
+        ok, reason = contractivity_ok(h)
+        if not ok:
+            halt_reason = f"contractivity_violated: {reason}"
+            log({"event": "contractivity_violated", "loop": t, "reason": reason})
+            break
+
+    # Coda — one emit regardless of T.
+    loop_summary = "\n".join(
+        f"loop {r['loop']}: specialist={r.get('specialist','?')} "
+        f"residual {r.get('residual_before','?')}->{r.get('residual_after','?')}"
+        f"{' CONVERGED' if r.get('converged') else ''}"
+        for r in trace
+        if "loop" in r
+    )
+    try:
+        text = coda_step(
+            e=e,
+            h=h,
+            loop_trace=loop_summary,
+            registry=registry,
+            policy=policy,
+            coda_role=coda_role,
+        )
+    except Exception as err:
+        text = f"(coda_error: {err}; final h_T summary: {h.summary or '(empty)'})"
+        halt_reason = f"{halt_reason}+coda_error"
+
+    total_ms = int((time.monotonic() - t0) * 1000)
+    log({
+        "event": "loop_end",
+        "loops_run": len(trace),
+        "halt_reason": halt_reason,
+        "total_ms": total_ms,
+        "residual_final": residual_magnitude(h),
+    })
+
+    return LoopResult(
+        text=text,
+        h_final=h,
+        loops_run=len([r for r in trace if "loop" in r]),
+        halt_reason=halt_reason,
+        trace=trace,
+    )
+
+# Recurrent probe CLI.
+
+def _probe_load_policy(policy_path: str | None = None):
+    try:
+        from .policy import load_policy
+    except ImportError:  # direct harness import under spark/ path
+        from harness.policy import load_policy  # type: ignore
+    return load_policy(policy_path) if policy_path else load_policy()
+
+
+def _probe_registry():
+    try:
+        from .providers import ProviderRegistry
+    except ImportError:
+        from harness.providers import ProviderRegistry  # type: ignore
+    return ProviderRegistry()
+
+
+def run_recurrent_probe_one(
+    prompt: str,
+    *,
+    registry: Any,
+    policy: Any,
+    max_loop_iters: int,
+    label: str,
+    out_path: Path | None,
+) -> dict[str, Any]:
+    """Run one T value through the recurrent loop and optionally append JSONL."""
+    t0 = time.monotonic()
+    events: list[dict[str, Any]] = []
+    result = run_recurrent_loop(
+        e=prompt,
+        registry=registry,
+        policy=policy,
+        max_loop_iters=max_loop_iters,
+        logger=events.append,
+    )
+    elapsed_ms = int((time.monotonic() - t0) * 1000)
+    record = {
+        "label": label,
+        "prompt": prompt[:400],
+        "max_loop_iters": max_loop_iters,
+        "loops_run": result.loops_run,
+        "halt_reason": result.halt_reason,
+        "elapsed_ms": elapsed_ms,
+        "residual_final": residual_magnitude(result.h_final),
+        "n_hypotheses_final": len(result.h_final.hypotheses),
+        "n_resolved_final": len(result.h_final.resolved),
+        "coda_text": result.text,
+        "trace": result.trace,
+    }
+    if out_path is not None:
+        out_path.parent.mkdir(parents=True, exist_ok=True)
+        with open(out_path, "a", encoding="utf-8") as f:
+            f.write(json.dumps(record, default=str) + "\n")
+    return record
+
+
+def recurrent_probe_main(argv: list[str] | None = None) -> int:
+    """Compare T=1 against deeper recurrent loop depths on stored prompts."""
+    import argparse
+    import os
+
+    ap = argparse.ArgumentParser(
+        description=(
+            "Probe the looped-orchestrate recurrent prototype. "
+            "Use --probe with --prompt or --prompts-file."
+        )
+    )
+    ap.add_argument("--probe", action="store_true", help="Run the recurrent probe CLI.")
+    src = ap.add_mutually_exclusive_group(required=False)
+    src.add_argument("--prompt", type=str, help="Single prompt to run.")
+    src.add_argument("--prompts-file", type=str, help="File with one prompt per line; blank lines/# ignored.")
+    ap.add_argument("--out", type=str, default=os.path.expanduser("~/logs/recurrent_probe.jsonl"))
+    ap.add_argument("--t-values", type=str, default="1,4")
+    ap.add_argument("--policy", type=str, default=None)
+    ap.add_argument("--quiet", action="store_true")
+    args = ap.parse_args(argv)
+
+    if not args.probe:
+        ap.print_help()
+        return 0
+    if not args.prompt and not args.prompts_file:
+        ap.error("--probe requires --prompt or --prompts-file")
+
+    policy = _probe_load_policy(args.policy)
+    registry = _probe_registry()
+
+    prompts: list[str] = []
+    if args.prompt:
+        prompts.append(args.prompt)
+    else:
+        text = Path(args.prompts_file).read_text(encoding="utf-8")
+        prompts = [line.strip() for line in text.splitlines() if line.strip() and not line.strip().startswith("#")]
+
+    try:
+        t_values = [int(x.strip()) for x in args.t_values.split(",") if x.strip()]
+    except ValueError:
+        print(f"bad --t-values: {args.t_values!r}", file=sys.stderr)
+        return 2
+
+    out_path = Path(args.out).expanduser()
+    for i, prompt in enumerate(prompts):
+        print(f"\n=== probe {i+1}/{len(prompts)}: {prompt[:80]!r} ===")
+        results_by_t: dict[int, dict[str, Any]] = {}
+        for T in t_values:
+            label = f"T={T}"
+            print(f"\n  running {label}...")
+            rec = run_recurrent_probe_one(
+                prompt,
+                registry=registry,
+                policy=policy,
+                max_loop_iters=T,
+                label=label,
+                out_path=out_path,
+            )
+            results_by_t[T] = rec
+            print(
+                f"  {label}: loops_run={rec['loops_run']} "
+                f"halt={rec['halt_reason']} residual={rec['residual_final']} "
+                f"elapsed={rec['elapsed_ms']}ms"
+            )
+            if not args.quiet:
+                print(f"  ---- coda ({label}) ----")
+                print(rec["coda_text"])
+                print(f"  ---- end coda ({label}) ----")
+        if len(t_values) >= 2:
+            baseline = results_by_t[t_values[0]]
+            deepest = results_by_t[t_values[-1]]
+            print(
+                f"\n  comparison (T={t_values[0]} -> T={t_values[-1]}): "
+                f"Δresidual={baseline['residual_final'] - deepest['residual_final']} "
+                f"(positive = deeper loop resolved more), "
+                f"Δelapsed={deepest['elapsed_ms'] - baseline['elapsed_ms']}ms"
+            )
+    print(f"\nwrote {out_path}")
+    return 0
+
+
+
+
+
+def _substrate_cli_main(argv: list[str] | None = None) -> int:
+    if argv and argv[0] == "--recurrent-probe":
+        return recurrent_probe_main(["--probe", *argv[1:]])
+    if argv and argv[0] == "--probe":
+        return recurrent_probe_main(["--probe", *argv[1:]])
+    print("substrate.py exposes prompt-building and recurrent probe helpers; use --recurrent-probe")
+    return 0
+
+
+if __name__ == "__main__":
+    raise SystemExit(_substrate_cli_main(sys.argv[1:]))
