+++ spark/tests/recurrent_probes.txt
@@ -1,4 +1,4 @@
-# Seed prompts for python3 -m spark.harness.recurrent --probe.
+# Seed prompts for python3 -m spark.harness.substrate --recurrent-probe.
 # Each line is one probe. Blank lines and # comments are ignored.
 # The point: prompts where a single pass is expected to be insufficient
 # (multi-hop reasoning, latent-dependent constraints, verification work)
@@ -16,4 +16,4 @@ A proposed harness change: add delegate_cb recursion to the chat role so bare gr
 Sketch a DeepSeekMoE-style "shared expert" analogue for the Vybn harness's role roster. Name what the shared expert is, what the routed experts are, and what the top-K routing semantics mean across a multi-turn conversation. Then identify the one invariant this proposal would break.
 
 # Counterfactual reasoning
-If we moved the orchestrate role from Opus 4.6 to Sonnet 4.6, what would we expect to see in the router_policy.yaml heuristics that currently fire on fallthrough — and would the contractivity monitor in recurrent.py catch the regression?
+If we moved the orchestrate role from Opus 4.6 to Sonnet 4.6, what would we expect to see in the router_policy.yaml heuristics that currently fire on fallthrough — and would the contractivity monitor in substrate.py catch the regression?
